[NuGet] Show license metadata in Manage Packages dialog
If the NuGet package source has license metadata then show this
information instead of the View License link. License metadata
can either be an license expression or a file.

If the license is a file then clicking View License will open a separate
dialog with the license read from the file.

Fixes VSTS #1005394 - Support the new "license" properties in the NuGet
package manager UI