[NuGet] Right align license metadata and show warnings as tooltip
Manage NuGet packages dialog changes:

Show license expression on the same line as the License label in the
Manage NuGet packages dialog.

Right align and wrap the license expression.

Move the license warnings to a tooltip for the warning icon instead
of showing the text directly in the dialog.

Move the warning icon so it is on the right hand side of the license
expression text.

License acceptance dialog changes:

Show license warnings in the warning icon tooltip instead of directly
on the dialog.

Show license warning icon on the same line as the license expression,
on the left hand side of the expression.

Fixes VSTS #1012071 - Right-align licenses to improve readability and
text wrapping

Fixes VSTS #1012079 - Use tooltip on warning icon to provide more
details