[NuGet] Ensure the file license dialog looks the same when displayed
Showing the file license dialog from the License Acceptance dialog
was showing the dialog with the native Xamarin.Mac toolkit but the
Manage NuGet packages dialog was using the non-native toolkit (GTK#).
Now the file license dialog is always launched with the native toolkit.
Also removed the horizontal scrollbar since this seems to cause the
text to unwrap using the native window.