[NuGet] Fix warning logged when clicking file license link
Clicking the View License link in the License Acceptance dialog was
logging a warning about failing to start the url. The problem was that
the custom url used to open the file license was being handled by the
default LinkLabel implementation. Now the NavigateToUrlEventArgs is
set as handled so this does not happen.